* Fixed compiler error & warning
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1483 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ npc.cpp
@@ -2067,7 +2067,7 @@ const NpcResponse* Npc::getResponse(const ResponseList& list, const Player* play
 
 			if(hasBitSet(RESPOND_LOWMONEY, params))
 			{
-				if(g_game.getMoney(player) >= (unsigned)npcState->price)
+				if((signed)g_game.getMoney(player) >= npcState->price)
 					continue;
 
 				++matchCount;
@@ -2075,7 +2075,7 @@ const NpcResponse* Npc::getResponse(const ResponseList& list, const Player* play
 
 			if(hasBitSet(RESPOND_LOWAMOUNT, params) || hasBitSet(RESPOND_NOAMOUNT, params))
 			{
-				if(player->__getItemTypeCount(npcState->itemId) >= (unsigned)npcState->amount)
+				if((signed)player->__getItemTypeCount(npcState->itemId) >= npcState->amount)
 					continue;
 
 				if(hasBitSet(RESPOND_LOWAMOUNT, params))
